The code was getting messy, this is the first stage of refactoring the code into a manageable smaller chunks.
