+++ plugins/DomainEnum/BingDomainEnum.py
@@ -1,0 +1,54 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+import re
+import urlparse
+from ..processor import enumratorBaseThreaded
+from ..layout import *
+
+#Import Colour Scheme
+G,Y,B,R,W = colour()
+
+class DomainSearch(enumratorBaseThreaded):
+    def __init__(self, domain, subdomains=None, q=None):
+        global verbose
+        verbose = subdomains
+        subdomains = subdomains or []
+        base_url = 'https://www.bing.com/search?q={query}&go=Submit&first={page_no}'
+        self.engine_name = "Bing"
+        self.MAX_DOMAINS = 30
+        self.MAX_PAGES = 0
+        enumratorBaseThreaded.__init__(self, base_url, self.engine_name,domain, subdomains,q=q)
+        self.q = q
+        return
+
+    def extract_domains(self, resp):
+        link_regx = re.compile('<li class="b_algo"><h2><a href="(.*?)"')
+        link_regx2 = re.compile('<div class="b_title"><h2><a href="(.*?)"')
+        try:
+            links = link_regx.findall(resp)
+            links2 = link_regx2.findall(resp)
+            links_list = links+links2
+
+            for link in links_list:
+                link = re.sub('<(\/)?strong>|<span.*?>|<|>', '', link)
+                if not link.startswith('http'):
+                    link="http://"+link
+                subdomain = urlparse.urlparse(link).netloc
+                if subdomain not in self.subdomains and subdomain != self.domain:
+                    if verbose:
+                        print "%s%s: %s%s"%(R, self.engine_name, W, subdomain)
+                    self.subdomains.append(subdomain.strip())
+        except Exception as e:
+            pass
+
+        return links_list
+
+    def generate_query(self):
+        if self.subdomains:
+            fmt = 'domain:{domain} -www.{domain} -{found}'
+            found = ' -'.join(self.subdomains[:self.MAX_DOMAINS])
+            query = fmt.format(domain=self.domain, found=found)
+        else:
+            query = "domain:{domain} -www.{domain}".format(domain=self.domain)
+        return query
