+++ plugins/DomainEnum/DNSdumpsterDomainEnum.py
@@ -1,0 +1,123 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+import multiprocessing
+import threading
+import requests
+import re
+import urlparse
+import dns.resolver
+
+from ..layout import *
+
+#Import Colour Scheme
+G,Y,B,R,W = colour()
+
+class DomainSearch(multiprocessing.Process):
+    def __init__(self, domain, subdomains=None, q=None, lock=threading.Lock()):
+        global verbose
+        verbose = subdomains
+        subdomains = subdomains or []
+        self.base_url = 'https://dnsdumpster.com/'
+        self.domain = urlparse.urlparse(domain).netloc
+        self.subdomains = []
+        self.live_subdomains = []
+        self.session = requests.Session()
+        self.engine_name = "DNSdumpster"
+        multiprocessing.Process.__init__(self)
+        self.threads = 70
+        self.lock = threading.BoundedSemaphore(value=self.threads)
+        self.q = q
+        self.timeout = 25
+        self.print_banner()
+        return
+
+    def run(self):
+        domain_list = self.enumerate()
+        for domain in domain_list:
+            self.q.append(domain)
+        return
+
+    def print_banner(self):
+        print G+"[-] Searching now in %s.." %(self.engine_name)+W
+        return
+
+    def check_host(self,host):
+        is_valid = False
+        Resolver = dns.resolver.Resolver()
+        Resolver.nameservers = ['8.8.8.8', '8.8.4.4']
+        self.lock.acquire()
+        try:
+            ip = Resolver.query(host, 'A')[0].to_text()
+            if ip:
+                if verbose:
+                    print "%s%s: %s%s"%(R, self.engine_name, W, host)
+                is_valid = True
+                self.live_subdomains.append(host)
+        except:
+            pass
+        self.lock.release()
+        return is_valid
+
+    def req(self, req_method, url, params=None):
+        params = params or {}
+        headers = {'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:38.0) Gecko/20100101 Firefox/40.0',
+        'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
+        'Accept-Language': 'en-GB,en;q=0.5',
+        'Accept-Encoding': 'gzip, deflate',
+        'Referer': 'https://dnsdumpster.com'
+        }
+
+        try:
+            if req_method == 'GET':
+                resp = self.session.get(url, headers=headers, timeout=self.timeout)
+            else:
+                resp = self.session.post(url, data=params, headers=headers, timeout=self.timeout)
+        except Exception as e:
+            print e
+            resp = None
+        return self.get_response(resp)
+
+    def get_response(self,response):
+    	if response is None:
+    		return 0
+        if hasattr(response, "text"):
+            return response.text
+        else:
+            return response.content
+
+    def get_csrftoken(self, resp):
+        csrf_regex = re.compile("<input type='hidden' name='csrfmiddlewaretoken' value='(.*?)' />",re.S)
+        token = csrf_regex.findall(resp)[0]
+        return token.strip()
+
+    def enumerate(self):
+        resp = self.req('GET', self.base_url)
+        token = self.get_csrftoken(resp)
+        params = {'csrfmiddlewaretoken':token, 'targetip':self.domain}
+        post_resp = self.req('POST', self.base_url, params)
+        self.extract_domains(post_resp)
+        for subdomain in self.subdomains:
+            t = threading.Thread(target=self.check_host,args=(subdomain,))
+            t.start()
+            t.join()
+        return self.live_subdomains
+
+
+    def extract_domains(self, resp):
+        tbl_regex = re.compile('<a name="hostanchor"><\/a>Host Records.*?<table.*?>(.*?)</table>',re.S)
+        link_regex = re.compile('<td class="col-md-4">(.*?)<br>',re.S)
+        links = []
+        try:
+            results_tbl = tbl_regex.findall(resp)[0]
+        except IndexError:
+            results_tbl = ''
+        links_list = link_regex.findall(results_tbl)
+        links = list(set(links_list))
+        for link in links:
+            subdomain = link.strip()
+            if not subdomain.endswith(self.domain):
+                continue
+            if subdomain and subdomain not in self.subdomains and subdomain != self.domain:
+                self.subdomains.append(subdomain)
+        return links
