+++ plugins/DomainEnum/GoogleDomainEnum.py
@@ -1,0 +1,61 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+import re
+import time
+
+from ..processor import enumratorBaseThreaded
+from ..layout import *
+
+#Import Colour Scheme
+G,Y,B,R,W = colour()
+
+class DomainSearch(enumratorBaseThreaded):
+    def __init__(self, domain, subdomains=None, q=None):
+        global verbose
+        verbose = subdomains
+        subdomains = subdomains or []
+        base_url = "https://google.com/search?q={query}&btnG=Search&hl=en-US&biw=&bih=&gbv=1&start={page_no}&filter=0"
+        self.engine_name = "Google"
+        self.MAX_DOMAINS = 11
+        self.MAX_PAGES = 200
+        super(DomainSearch, self).__init__(base_url, self.engine_name, domain, subdomains, q=q)
+        self.q = q
+        return
+
+    def extract_domains(self, resp):
+        link_regx = re.compile('<cite.*?>(.*?)<\/cite>')
+        try:
+            links_list = link_regx.findall(resp)
+            for link in links_list:
+                link = re.sub('<span.*>', '', link)
+                if not link.startswith('http'):
+                    link="http://"+link
+                subdomain = urlparse.urlparse(link).netloc
+                if subdomain and subdomain not in self.subdomains and subdomain != self.domain:
+                    if verbose:
+                        print "%s%s: %s%s"%(R, self.engine_name, W, subdomain)
+                    self.subdomains.append(subdomain)
+        except Exception as e:
+            pass
+        return links_list
+
+    def check_response_errors(self, resp):
+        if 'Our systems have detected unusual traffic' in resp:
+            print R+"[!] Error: Google probably now is blocking our requests"+W
+            print R+"[~] Finished now the Google Enumeration ..."+W
+            return False
+        return True
+
+    def should_sleep(self):
+        time.sleep(5)
+        return
+
+    def generate_query(self):
+        if self.subdomains:
+            fmt = 'site:{domain} -www.{domain} -{found}'
+            found = ' -'.join(self.subdomains[:self.MAX_DOMAINS-2])
+            query = fmt.format(domain=self.domain, found=found)
+        else:
+            query = "site:{domain} -www.{domain}".format(domain=self.domain)
+        return query
