+++ plugins/DomainEnum/YahooDomainEnum.py
@@ -1,0 +1,62 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+import re
+import urlparse
+
+from ..processor import enumratorBaseThreaded
+from ..layout import *
+
+#Import Colour Scheme
+G,Y,B,R,W = colour()
+
+class DomainSearch(enumratorBaseThreaded):
+    def __init__(self, domain, subdomains=None, q=None):
+        global verbose
+        verbose = subdomains
+        subdomains = subdomains or []
+        base_url = "https://search.yahoo.com/search?p={query}&b={page_no}"
+        self.engine_name = "Yahoo"
+        self.MAX_DOMAINS = 10
+        self.MAX_PAGES = 0
+        super(DomainSearch, self).__init__(base_url, self.engine_name,domain, subdomains, q=q)
+        self.q = q
+        return
+
+    def extract_domains(self, resp):
+        link_regx2 = re.compile('<span class=" fz-15px fw-m fc-12th wr-bw.*?">(.*?)</span>')
+        link_regx = re.compile('<span class="txt"><span class=" cite fw-xl fz-15px">(.*?)</span>')
+        try:
+            links = link_regx.findall(resp)
+            links2 = link_regx2.findall(resp)
+            links_list = links+links2
+            for link in links_list:
+                link = re.sub("<(\/)?b>","", link)
+                if not link.startswith('http'):
+                    link="http://"+link
+                subdomain = urlparse.urlparse(link).netloc
+                if not subdomain.endswith(self.domain):
+                    continue
+                if subdomain and subdomain not in self.subdomains and subdomain != self.domain:
+                    if verbose:
+                        print "%s%s: %s%s"%(R, self.engine_name, W, subdomain)
+                    self.subdomains.append(subdomain)
+        except Exception as e:
+            pass
+
+        return links_list
+
+    def should_sleep(self):
+        return
+
+    def get_page(self,num):
+        return num + 10
+
+    def generate_query(self):
+        if self.subdomains:
+            fmt = 'site:{domain} -domain:www.{domain} -domain:{found}'
+            found = ' -domain:'.join(self.subdomains[:77])
+            query = fmt.format(domain=self.domain, found=found)
+        else:
+            query = "site:{domain}".format(domain=self.domain)
+        return query
